+++ src/animations/animateChildren/animateChildren.animation.ts
@@ -1,0 +1,9 @@
+import {animateChild, animation, AnimationReferenceMetadata, query} from '@angular/animations';
+
+/**
+ * Animates children animations
+ */
+export const animateChildrenAnimation: AnimationReferenceMetadata = animation(
+[
+    query('@*', animateChild(), {optional: true})
+]);