+++ src/index.ts
@@ -4,11 +4,13 @@ export {flyInAnimation, FlyInParams} from './animations/flyIn/flyIn.animation';
 export {flyOutAnimation, FlyOutParams} from './animations/flyOut/flyOut.animation';
 export {slideInAnimation, SlideInParams} from './animations/slideIn/slideIn.animation';
 export {slideOutAnimation, SlideOutParams} from './animations/slideOut/slideOut.animation';
+export {animateChildrenAnimation} from './animations/animateChildren/animateChildren.animation';
 export {fadeInOutTriggerFactory, fadeInOutTrigger} from './triggers/fadeInOut/fadeInOut.trigger';
 export {flyInTriggerFactory, flyInTrigger} from './triggers/flyIn/flyIn.trigger';
 export {flyInOutTriggerFactory, flyInOutTrigger} from './triggers/flyInOut/flyInOut.trigger';
 export {flyOutTriggerFactory, flyOutTrigger} from './triggers/flyOut/flyOut.trigger';
 export {slideInOutTriggerFactory, slideInOutTrigger} from './triggers/slideInOut/slideInOut.trigger';
 export {visibleFlyInOutTriggerFactory, visibleFlyInOutTrigger} from './triggers/visibleFlyInOut/visibleFlyInOut.trigger';
+export {animateChildrenTrigger} from './triggers/animateChildren/animateChildren.trigger';
 export {VISIBLE} from './triggers/stateNames';